+++ test/fixtures/if.rb
@@ -24,3 +24,7 @@ foo = (baz if bar)
 if foo += 1
   foo
 end
+%
+if (foo += 1)
+  foo
+end