+++ test/fixtures/if_mod.rb
@@ -8,3 +8,8 @@ if foo
 end
 %
 bar if foo # comment
+%
+foo = barrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrr if foo
+-
+foo =
+  barrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrr if foo