+++ test/fixtures/unless.rb
@@ -24,3 +24,7 @@ foo = (baz unless bar)
 unless foo += 1
   foo
 end
+%
+unless (foo += 1)
+  foo
+end